+++ src/configuration/utils.ts
@@ -1,0 +1,16 @@
+import { SyntheticEvent } from "react";
+
+import { SelectableValue } from "@grafana/data";
+
+
+export const getValueFromEventItem = (eventItem: SyntheticEvent<HTMLInputElement> | SelectableValue<string>) => {
+  if (!eventItem) {
+    return '';
+  }
+
+  if (eventItem.hasOwnProperty('currentTarget')) {
+    return eventItem.currentTarget.value;
+  }
+
+  return (eventItem as SelectableValue<string>).value;
+};